+++ examples/wobbly_function/wobbly_function.py
@@ -2,6 +2,7 @@
 Usage:
     python3 function.py -filebasename <folder-to-store-output>/<somefilebasename> -x 3 -y 5.5
 """
+
 import csv
 from argparse import ArgumentParser
 from typing import Any