+++ tests/test_execution.py
@@ -1,4 +1,5 @@
 """Unit tests for the `yotse.execution` package's `Executor` class."""
+
 import json
 import math
 import os