+++ yotse/utils/utils.py
@@ -18,6 +18,7 @@ ndarray_to_list(numpy_array: np.ndarray) -> List[Any]:
 list_to_numpy_array(list_data: List[Any]) -> np.ndarray:
     Convert a Python list to a NumPy array.
 """
+
 import os
 from typing import Any
 from typing import List